refactor: add temperature rounding to display to the user

+++ src/main/java/com/projects/weather/dto/location/response/LocationWithWeatherDto.java
@@ -16,8 +16,8 @@ public record LocationWithWeatherDto(Long id,
     }
 
     @Builder
-    public record TemperatureInfo(Double temperature,
-                                   Double feelsLikeTemp,
-                                   Double humidity) {
+    public record TemperatureInfo(int temperature,
+                                   int feelsLikeTemp,
+                                   int humidity) {
     }
 }
